Refactored CMake, build.py and directory structure
* Improved directory structure (tests, example subdirectories)

* Refactored cmake files

* Refactored build.py

* Added multicore build option to build.py

* Removed deprecated windows store target for Windows < 10.0

* Updated dockerfiles

* Added file grouping

* Removed curl

+++ boden/include/bdn/StringBuffer.h
@@ -455,7 +455,7 @@ private:
 				{
 					size_t avail = chunk.used() - posInChunk;
 
-					if(offset>avail)
+					if((size_t)offset>avail)
 					{
 						offset-=avail;
 						++chunkIt;
@@ -479,7 +479,7 @@ private:
 				}
 				else
 				{
-					if( -offset > posInChunk )
+					if( ((size_t)-offset) > posInChunk )
 					{
 						offset += posInChunk;
